requirements.txt
# ERP-Assistent – Python Dependencies
# Installieren mit: pip install -r requirements.txt

# Web-Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# KI-Modelle
torch>=2.3.0
transformers>=4.45.0
accelerate>=0.33.0
bitsandbytes>=0.43.0    # 4-bit Quantisierung für große Modelle
Pillow>=10.0.0
sentencepiece>=0.2.0

# Excel
pandas>=2.2.0
openpyxl>=3.1.0
xlrd>=2.0.1

# GUI-Steuerung
pyautogui>=0.9.54
pygetwindow>=0.0.9
numpy>=1.26.0
